Explicitly specify 16 bit ints for audio output bits and analog readings.
This does not make any difference on AVR, but for 32bit boards such as
Teensy or STM32, "int" is actually 32bits. While this is not much of a
problem in most places, it causes excessive memory consumption in the
audio output buffer (and to a lesser degree in the analog readings buffer).

Unfortunately this will break existing sketches on 32 bit boards (but only
on these), as it changes the signature of updateAudio().

+++ MozziGuts.cpp
@@ -73,9 +73,9 @@ PWM frequency tests
 
 //-----------------------------------------------------------------------------------------------------------------
 // ring buffer for audio output
-CircularBuffer <unsigned int> output_buffer; // fixed size 256
+CircularBuffer <uint16_t> output_buffer; // fixed size 256
 #if (STEREO_HACK == true)
-CircularBuffer <unsigned int> output_buffer2; // fixed size 256
+CircularBuffer <uint16_t> output_buffer2; // fixed size 256
 #endif
 //-----------------------------------------------------------------------------------------------------------------
 
@@ -251,10 +251,10 @@ void audioHook() // 2us excluding updateAudio()
 	if (!output_buffer.isFull()) {
 		#if (STEREO_HACK == true)
 		updateAudio(); // in hacked version, this returns void
-		output_buffer.write((unsigned int) (audio_out_1 + AUDIO_BIAS));
-		output_buffer2.write((unsigned int) (audio_out_2 + AUDIO_BIAS));
+		output_buffer.write((uint16_t) (audio_out_1 + AUDIO_BIAS));
+		output_buffer2.write((uint16_t) (audio_out_2 + AUDIO_BIAS));
 		#else
-		output_buffer.write((unsigned int) (updateAudio() + AUDIO_BIAS));
+		output_buffer.write((uint16_t) (updateAudio() + AUDIO_BIAS));
 		#endif
 
 	}
